copy issue to ticket table

+++ lib-commons/src/main/java/dev/scaraz/mars/common/domain/request/CreateIssueDTO.java
@@ -3,13 +3,12 @@ package dev.scaraz.mars.common.domain.request;
 import dev.scaraz.mars.common.domain.response.IssueParamDTO;
 import dev.scaraz.mars.common.tools.enums.Product;
 import dev.scaraz.mars.common.tools.enums.Witel;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import jakarta.validation.constraints.NotNull;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -33,7 +32,7 @@ public class CreateIssueDTO {
     private String description;
 
     @NotNull
-    private BigDecimal score;
+    private Double score;
 
     @Builder.Default
     private List<IssueParamDTO> params = new ArrayList<>();
